moved SegmentChecker to separate file
